pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autogal"
version = "0.1.0"
description = "AI-driven Ren'Py game generation agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AutoGal Team" }
]
keywords = ["renpy", "visual-novel", "ai", "game-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "aiofiles>=23.0.0",
    "rembg>=2.0.50",    # 新增：用于自动去除图片背景
    "Pillow>=10.0.0",   # 新增：Python 图像处理基础库
    "onnxruntime>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
backend = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "celery>=5.3.0",
    "redis>=4.5.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
autogal = "autogal.__main__:main"

[project.urls]
Homepage = "https://github.com/yourname/autogal"
Documentation = "https://github.com/yourname/autogal#readme"
Repository = "https://github.com/yourname/autogal"

[tool.hatch.build.targets.wheel]
packages = ["src/autogal"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
